check if instance pooling is enabled in cache

+++ classes/Propel/Runtime/ActiveQuery/CachedInstancePoolTrait.php
@@ -3,6 +3,7 @@
 namespace Propel\Runtime\ActiveQuery;
 
 use Kopper\Cache\GlobalCache;
+use Propel\Runtime\Propel;
 
 trait CachedInstancePoolTrait {
 
@@ -21,26 +22,32 @@ trait CachedInstancePoolTrait {
   }
 
   public static function addInstanceToPool($object, $key = null) {
-    static::defaultAddInstanceToPool($object, $key);
-    GlobalCache::set(static::getCacheKey($object), $object);
+    if (Propel::isInstancePoolingEnabled()) {
+      static::defaultAddInstanceToPool($object, $key);
+      GlobalCache::set(static::getCacheKey($object), $object);
+    }
   }
 
   public static function removeInstanceFromPool($value) {
-    static::defaultRemoveInstanceFromPool($value);
-    GlobalCache::delete(static::getCacheKey($value));
+    if (Propel::isInstancePoolingEnabled()) {
+      static::defaultRemoveInstanceFromPool($value);
+      GlobalCache::delete(static::getCacheKey($value));
+    }
   }
 
   public static function getInstanceFromPool($key) {
     $object = static::defaultGetInstanceFromPool($key);
 
-    if (empty($object) === true) {
-      $object = GlobalCache::get(static::getPrefix() . $key);
-    }
+    if (Propel::isInstancePoolingEnabled()) {
+      if (empty($object) === true) {
+        $object = GlobalCache::get(static::getPrefix() . $key);
+      }
 
-    if ($object === false) {
-      $object = null;
-    } else {
-      static::defaultAddInstanceToPool($object, $key);
+      if ($object === false) {
+        $object = null;
+      } else {
+        static::defaultAddInstanceToPool($object, $key);
+      }
     }
 
     return $object;
